Move PreferencesDialog handling to presenter
