+++ de-lib/src/main/java/org/iplantc/de/desktop/client/DesktopView.java
@@ -228,6 +228,8 @@ public interface DesktopView extends IsWidget {
 
         void onCollaboratorsClick();
 
+        void onPreferencesClick();
+
         void onContactSupportClick();
 
         void onDocumentationClick();
